seeds: update the seeds list
The list of seeds that we currently have is really old
and also are really random.

What I did is to peak some of the most popular nodes (including some CLN
nodes) and update our list of seeds.

However, I am open to an objection is some of you want to keep ar add others
seeds nodes.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>